Separação de queries de entrada de moradores e visitantes
